requirements.txt
blinker==1.9.0
click==8.1.7
cloup==3.0.5
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
Flask==3.1.0
fonttools==4.55.3
fpdf==1.7.2
glcontext==3.0.0
isosurfaces==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
lxml==5.3.0
manim==0.18.1
ManimPango==0.6.0
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.0.3
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pycairo==1.27.0
pydub==0.25.1
pygame==2.6.1
pyglet==2.0.18
PyGLM==2.7.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-pptx==1.0.2
rich==13.9.4
scipy==1.14.1
screeninfo==0.8.1
setuptools==75.6.0
six==1.17.0
skia-pathops==0.8.0.post2
srt==3.5.3
svgelements==1.9.6
tk==0.1.0
tqdm==4.67.1
typing_extensions==4.12.2
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
XlsxWriter==3.2.0
